Regression

Description

The module implements some basic regression analysis methods such as least squares. Some important values about that the estimate are stored in the Plot data type automatically.

Types

data Correlation

Correlation describes the some of the most popular types of dependence.

Constructors

Constant

Linear

Squared

data Base a

Base contains the original data that a plot is based on.

Constructors

Base

Fields

ys :: Matrix a

The dependent variable data vector

xs :: Matrix a

The independent variable data matrix

data Plot a

Plot contains information about a regression estimate.

Constructors

Plot

Fields

coefs :: Matrix a

Estimate coefficients

values :: Matrix a

Estimated values for the given predictor values

residuals :: Matrix a

Estimation residuals

correlation :: Correlation

The used correlation for the plot

base :: Base a

The measured base for the regression

Instances

Utility functions

plot

Arguments

:: Num a

=> Matrix a
Regressor matrix (independent variables)

-> Matrix a
Dependent variable vector

-> Correlation
Correlation description type

-> Matrix a
Matrix for the estimated coefficients

-> Plot a
The plot containing infromation about the estimate

Constructor method for the Plot data type.

calculateFittedValues

Arguments

:: Num a

=> Matrix a
Matrix for the estimated coefficients

-> Matrix a
Regressor matrix (independent variables)

-> Matrix a
Vector of estimated values of the dependent variable

Calculates the fitted values based on an estimate.

calculateResiduals

Arguments

:: Num a

=> Matrix a
The dependent data variables

-> Matrix a
The estimated values

-> Matrix a
Matrix of the residuals (difference between the first two)

Calculates the residuals in an estimation.

(+) :: Correlation -> Correlation -> [Correlation]

Allows combinations of correlation types.

(.%.) :: Num a => Matrix a -> Matrix a -> Matrix a

A synonym for matrix multiplication.

Regression functions

leastSquaresEstimator

Arguments

:: (Ord a, Eq a, Fractional a)

=> Matrix a
Regressor matrix (independent variables)

-> Matrix a
Dependent variable vector

-> Maybe (Matrix a)
Maybe the estimated coefficient matrix

Calculates the least squares estimates for the data matrices if it is possible. Generally it is not possible if the columns of the regressor matrix aren't linearly independent and therefore its Gramian isn't invertible.

plotFromMatrix

Arguments

:: (Ord a, Eq a, Fractional a)

=> Correlation
Correlation descriptor

-> Matrix a
Regressor matrix (independent variables)

-> Matrix a
Dependent variable vector

-> Maybe (Plot a)
Maybe the estimated plot

Calculates a least squares Plot based on the data matrices of the dependent and independent variables and correlation information. This might fail due to linear dependence in the regressor matrix. Only supports linear correlation so far.

plotFromBase :: (Ord a, Eq a, Fractional a) => Correlation -> Base a -> Maybe (Plot a)