RegressionDiagnostics

Description

The module features some of the most popular regression diagnostics algorithms and calculations. Very little is presumed about the distribution of the source data.

Statistical utilities

quantile :: (Ord a, RealFrac a, Floating a) => a -> [a] -> a

Permutation utilities

permuteBase

Arguments

:: Floating a

=> Base a
The base that will be permuted

-> [Int]
The target columns in the data matrix (xs)

-> Permute
The latest permute

-> (Permute, Base a)
The latest permute and the permutated base

Permutates the data matrix (xs) of a base of a Plot.

permuteMatrix

Arguments

:: Floating a

=> Matrix a
The target matrix

-> [Int]
The columns to permute

-> Permute
The permutation to implement

-> (Permute, Matrix a)
The permuted matrix

Permutates column of a matrix. This is ugly. Ugh...

permutationTest

Arguments

:: Floating a

=> (Plot a -> a)
Test function (such as rsquared)

-> (Base a -> Maybe (Plot a))
Plotter function

-> Base a
The base to permute (should be more specific)

-> [Int]
Target columns in the base matrix (xs)

-> Int
Amount of permutations (n)

-> Permute
The initial permutation

-> [Maybe a]
Results of the tester function in n permutations

A generic permutation test function. It needs a tester and plotter functions to test the quality of new plots and to generate them from the permuted Base data. So far the test can only permute columns and requires a manual insertion of a Permute.

permutationTest' :: Floating a => (Plot a -> a) -> (Base a -> Maybe (Plot a)) -> Base a -> [Int] -> Int -> Permute -> [Maybe a] -> [Maybe a]

Helper function for the actual permutationTest.

Test functions

rsquared :: Floating a => Plot a -> a

Permutation Tests

regressionCoefSignificance

Arguments

:: (Ord a, RealFrac a, Floating a)

=> (Base a -> Maybe (Plot a))
Plotter function (TODO: make this a type)

-> Base a
The base to permute

-> Int
The coefficient index

-> Int
The amount of permutations

-> Permute
Initial permutation

-> a
Confidence level (alpha)

-> a
1 - alpha quantile of the R-squared list

Function approximates the significance of a regression coefficient. It takes a confidence level alpha that is used to calculate a quantile of the permutation results.

coefResults

Arguments

:: (Ord a, Floating a)

=> (Base a -> Maybe (Plot a))
Plotter function (TODO: make this a type)

-> Base a
The base to permute

-> Int
The coefficient index

-> Int
The amount of permutations

-> Permute
Initial permutation

-> [a]

Debug function of regressionCoefSignificance.